backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
typing_extensions
supabase
openai
openai-agents
